% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{DEoptim_metaRVM}
\alias{DEoptim_metaRVM}
\title{Calibrate MetaRVM using DEoptim}
\usage{
DEoptim_metaRVM(config, ground_truth, params_to_infer, deoptim_settings)
}
\description{
Wrapper for \code{DEoptim} to calibrate MetaRVM.
}
\keyword{internal}
