% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R, R/parse_config.R
\name{draw_sample}
\alias{draw_sample}
\title{Title}
\usage{
draw_sample(config_list, N_pop, seed = NULL)

draw_sample(config_list, N_pop, seed = NULL)
}
\arguments{
\item{config_list}{A list of configurations}

\item{N_pop}{Number of subpopulations}
}
\value{
A random sample drawn from the distribution specified by the dist component

A random sample drawn from the distribution specified by the dist component
}
\description{
Title

Function to draw samples for distributional parameters
}
