% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_output}
\alias{format_output}
\title{Format simulation output from wide to tall}
\usage{
format_output(out)
}
\arguments{
\item{out}{output from the meta_sim function with added "rep" column}
}
\value{
a tall output table
}
\description{
Format simulation output from wide to tall
}
