% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_disease_state}
\alias{get_disease_state}
\title{Extract disease states}
\usage{
get_disease_state(
  out,
  disease_states = c("S", "E", "H", "D", "I_presymp", "I_asymp", "I_symp", "R", "V")
)
}
\arguments{
\item{out}{A matrix of output from \code{meta_sim} function.}

\item{disease_states}{A character scalar or vector of disease states. The
disease state names should match with \code{c("S", "E", "H", "D", "I_presymp", "I_asymp", "I_symp", "R", "V")}}
}
\value{
A long tibble
}
\description{
Extract disease states
}
\examples{
\dontrun{
get_disease_state(out, disease_states = c("I_presymp", "I_asymp", "I_symp"))
}
}
