% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_HCEZ}
\alias{get_HCEZ}
\title{Extract disease states by HCE zone}
\usage{
get_HCEZ(
  out,
  HCEZ_id = 1:6,
  disease_states = c("S", "E", "H", "D", "I_presymp", "I_asymp", "I_symp", "R", "V")
)
}
\arguments{
\item{out}{matrix of output from \code{meta_sim} function.}

\item{HCEZ_id}{A scalar or vector of HCE zone ids.}

\item{disease_states}{A character scalar or vector of disease states. The
disease state names should match with \code{c("S", "E", "H", "D", "I_presymp", "I_asymp", "I_symp", "R", "V")}}
}
\value{
A long tibble
}
\description{
Extract disease states by HCE zone
}
\examples{
\dontrun{
get_HCEZ(out, disease_states = c("I_presymp", "I_asymp", "I_symp"), HCEZ_id = c(2, 4))
}
}
