% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaODIN.R
\name{meta_sim}
\alias{meta_sim}
\title{Metapopulation respiratory virus model Simulator}
\usage{
meta_sim(
  N_pop,
  ts,
  tv,
  S0,
  I0,
  P0,
  R0,
  H0 = rep(0, N_pop),
  D0 = rep(0, N_pop),
  Ia0 = rep(0, N_pop),
  Ip0 = rep(0, N_pop),
  E0 = rep(0, N_pop),
  m_weekday_day,
  m_weekday_night,
  m_weekend_day,
  m_weekend_night,
  delta_t,
  vac_mat,
  dv,
  de,
  pea,
  dp,
  da,
  ds,
  psr,
  dh,
  phr,
  dr,
  ve,
  nsteps,
  is.stoch = FALSE,
  seed = NULL,
  do_chk = FALSE,
  chk_file_name = NULL
)
}
\arguments{
\item{N_pop}{Number of sub-populations}

\item{ts}{A scalar or vector, transmission rate for susceptible population}

\item{tv}{A scalar or vector, transmission rate for vaccinated population}

\item{S0}{A vector of initial number of susceptible people in N_pop subpopulations}

\item{I0}{A vector of initial number of infected people in N_pop subpopulations}

\item{P0}{A vector of population sizes for N_pop subpopulations}

\item{R0}{A vector of initial number of recovered people in N_pop subpopulations}

\item{H0}{A vector of initial number of hospitalized people in N_pop subpopulations}

\item{D0}{A vector of initial number of dead people in N_pop subpopulations}

\item{Ia0}{A vector of initial number of Infectious asymptomatic people in N_pop subpopulations}

\item{Ip0}{A vector of initial number of Infectious presymptomatic people in N_pop subpopulations}

\item{E0}{A vector of initial number of Exposed people in N_pop subpopulations}

\item{m_weekday_day}{mixing matrix for weekday 6 am - 6 pm}

\item{m_weekday_night}{mixing matrix for weekday 6 pm - 6 am}

\item{m_weekend_day}{mixing matrix for weekend 6 am - 6 pm}

\item{m_weekend_night}{mixing matrix for weekend 6 pm - 6 am}

\item{delta_t}{A positive real number indicating the discrete time increment}

\item{vac_mat}{A matrix of order nsteps x (1 + N_pop), first column in time indices
and the rest are vaccination counts in N_pop strata.}

\item{dv}{mean number of days in Vaccinated state before waning immunity completely}

\item{de}{mean number of days in Exposed state}

\item{pea}{Proportion of people becoming infectious asymptomatic from exposed}

\item{dp}{mean number of days in Infectious presymptomatic state}

\item{da}{mean number of days in Infectious presymptomatic state}

\item{ds}{mean number of days in Infectious symptomatic state}

\item{psr}{Proportion of people recovered from infectious symptomatic}

\item{dh}{mean number of days in Hospitalized state}

\item{phr}{Proportion of people Recovered from Hospitalized}

\item{dr}{mean number of days in Recovered state}

\item{ve}{Vaccination efficacy, a value between 0 and 1}

\item{nsteps}{Number of discrete evolution in the simulation}

\item{is.stoch}{1 if the simulation is stochastic, 0 otherwise}

\item{seed}{optional, for reproducibility, only application when is.stoch = 1}

\item{do_chk}{boolean, if the model needs to be checkpointed}

\item{chk_file_name}{if do_chk is TRUE, a file name}
}
\value{
A data.table of model output in long format
}
\description{
The main simulator which compiles and run an ODIN model in the background
}
