% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaRVM-package.R
\docType{package}
\name{MetaRVM-package}
\alias{MetaRVM}
\alias{MetaRVM-package}
\title{MetaRVM: Meta-population compartmental model for respiratory virus diseases}
\description{
Simulation of meta-population compartmental model
}
\author{
\strong{Maintainer}: Arindam Fadikar \email{afadikar@anl.gov}

}
\keyword{internal}
