% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MetaRVMCheck}
\alias{MetaRVMCheck}
\title{MetaRVM Checkpoint Class}
\description{
R6 class that stores a saved simulation state produced by \code{\link[=meta_sim]{meta_sim()}}.
Used to resume or branch a simulation from an intermediate time point.
}
\details{
Checkpoint objects are written automatically during a simulation run when
\code{do_chk = TRUE} is set in the simulation config. Each checkpoint captures
the full compartment state (S, E, I, H, R, D, V, …) and all disease
parameters at the saved time step, enabling a resumed run to pick up exactly
where it left off.

\code{MetaRVMCheck} extends \code{\link{MetaRVMConfig}}: all \verb{$get()}, \verb{$get_category_names()},
and related accessor methods are inherited and work identically. Required
checkpoint fields are disease-specific and validated against the disease
registry at construction time.
}
\examples{
\dontrun{
# Checkpoints are created automatically when do_chk = TRUE in the config.
# To inspect a saved checkpoint:
chk_data <- readRDS("my_checkpoint.rds")
chk <- MetaRVMCheck$new(chk_data)
chk$get("chk_time_step")
chk$get("N_pop")
}

}
\seealso{
\code{\link[=metaRVM]{metaRVM()}} for enabling checkpointing via the \code{simulation_config}
section; \code{\link{MetaRVMConfig}} for the parent class and its accessor methods.
}
\author{
Arindam Fadikar
}
\section{Super class}{
\code{\link[MetaRVM:MetaRVMConfig]{MetaRVM::MetaRVMConfig}} -> \code{MetaRVMCheck}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{check_data}}{List containing all parsed checkpoint data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetaRVMCheck-new}{\code{MetaRVMCheck$new()}}
\item \href{#method-MetaRVMCheck-clone}{\code{MetaRVMCheck$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="get"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-get'><code>MetaRVM::MetaRVMConfig$get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="get_all"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-get_all'><code>MetaRVM::MetaRVMConfig$get_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="get_all_categories"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-get_all_categories'><code>MetaRVM::MetaRVMConfig$get_all_categories()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="get_category_names"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-get_category_names'><code>MetaRVM::MetaRVMConfig$get_category_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="get_category_values"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-get_category_values'><code>MetaRVM::MetaRVMConfig$get_category_values()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="get_pop_map"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-get_pop_map'><code>MetaRVM::MetaRVMConfig$get_pop_map()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="list_parameters"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-list_parameters'><code>MetaRVM::MetaRVMConfig$list_parameters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="parameter_summary"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-parameter_summary'><code>MetaRVM::MetaRVMConfig$parameter_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="print"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-print'><code>MetaRVM::MetaRVMConfig$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="MetaRVM" data-topic="MetaRVMConfig" data-id="set"><a href='../../MetaRVM/html/MetaRVMConfig.html#method-MetaRVMConfig-set'><code>MetaRVM::MetaRVMConfig$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaRVMCheck-new"></a>}}
\if{latex}{\out{\hypertarget{method-MetaRVMCheck-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MetaRVMCheck object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaRVMCheck$new(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{A list containing checkpoint data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MetaRVMCheck} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaRVMCheck-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MetaRVMCheck-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaRVMCheck$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
