% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optim_metaRVM}
\alias{optim_metaRVM}
\title{Calibrate MetaRVM using optim}
\usage{
optim_metaRVM(config, ground_truth, params_to_infer, optim_settings)
}
\description{
Wrapper for \code{optim} to calibrate MetaRVM.
}
\keyword{internal}
