% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{parse_config}
\alias{parse_config}
\title{Function to read and parse model inputs from yaml file}
\usage{
parse_config(config_file)
}
\arguments{
\item{config_file}{A yaml configuration file.}
}
\value{
a named list of (key, value) pairs of the parameters
}
\description{
Function to read and parse model inputs from yaml file
}
\details{
Additional details...
}
