% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metaRVM}
\alias{plot_metaRVM}
\title{Aggregate outcome plot}
\usage{
plot_metaRVM(long_out_daily, conf_level, value_column, is.plotly = TRUE)
}
\arguments{
\item{is.plotly}{}
}
\description{
Aggregate outcome plot
}
