% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metaRVM_by_cateory}
\alias{plot_metaRVM_by_cateory}
\title{Title}
\usage{
plot_metaRVM_by_cateory(
  long_out,
  conf_level,
  pop_map_df,
  start_date,
  by_category = NULL
)
}
\arguments{
\item{by_category}{}
}
\description{
Title
}
