% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metaRVM_by_cateory_by_compartment}
\alias{plot_metaRVM_by_cateory_by_compartment}
\title{Title}
\usage{
plot_metaRVM_by_cateory_by_compartment(
  long_out,
  conf_level,
  pop_map_df,
  start_date,
  by_category = NULL,
  gg_theme = NULL
)
}
\arguments{
\item{gg_theme}{}
}
\description{
Title
}
