% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_simout}
\alias{subset_simout}
\title{Title}
\usage{
subset_simout(
  long_out,
  start_date,
  pop_map_df,
  agecats,
  racecats,
  zones,
  disease_states
)
}
\arguments{
\item{disease_states}{}
}
\description{
Title
}
