% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaODIN.R
\name{meta_sim}
\alias{meta_sim}
\title{Metapopulation respiratory virus model Simulator}
\usage{
meta_sim(
  N_pop,
  beta_i,
  beta_v,
  S0,
  I0,
  P0,
  V0,
  m_weekday_day,
  m_weekday_night,
  m_weekend_day,
  m_weekend_night,
  delta_t,
  tvac,
  vac_mat,
  VtoS,
  EtoIpresymp,
  etopa,
  pretoIsymp,
  IasymptoR,
  IsymptoRH,
  istohr,
  HtoRD,
  htor,
  RtoS,
  vac_eff,
  nsteps,
  is.stoch = FALSE,
  seed = NULL
)
}
\arguments{
\item{N_pop}{Number of sub-populations}

\item{beta_i}{Rate of infection, value between 0 and 1}

\item{beta_v}{Rate of infection for vaccinated people, value between 0 and 1}

\item{S0}{A vector of initial number of susceptible people in N_pop subpopulations}

\item{I0}{A vector of initial number of infected people in N_pop subpopulations}

\item{P0}{A vector of population sizes for N_pop subpopulations}

\item{V0}{A vector of initial number of vaccinated people in N_pop subpopulations}

\item{m_weekday_day}{mixing matrix for weekday 6 am - 6 pm}

\item{m_weekday_night}{mixing matrix for weekday 6 pm - 6 am}

\item{m_weekend_day}{mixing matrix for weekend 6 am - 6 pm}

\item{m_weekend_night}{mixing matrix for weekend 6 pm - 6 am}

\item{delta_t}{A positive real number indicating the discrete time increment}

\item{tvac}{A vector of time indices for vaccination}

\item{vac_mat}{A matrix of order (length of tvac) x N_pop}

\item{VtoS}{Rate of transitioning from vaccinated to susceptible}

\item{EtoIpresymp}{rate of transitioning from Exposed to infectious presymptomatic or asymptomatic}

\item{etopa}{Proportion of people becoming infectious presymptomatic from exposed}

\item{pretoIsymp}{Rate of transitioning from infectious presymptomatic to infectious symptomatic}

\item{IasymptoR}{Rate of transitioning from infectious presymptomatic to recovered}

\item{IsymptoRH}{Rate for transitioning from infectious symptomatic to recovered or hospitalized}

\item{istohr}{Proportion of people hospitalized from infectious symptomatic}

\item{HtoRD}{Rate for transitioning from hospitalized to recovered or dead}

\item{htor}{Proportion of people recovered from hospitalized}

\item{RtoS}{Rate for transitioning from recovered to susceptible}

\item{vac_eff}{Vaccination efficacy, a value between 0 and 1}

\item{nsteps}{Number of discrete evolution in the simulation}

\item{is.stoch}{1 if the simulation is stochastic, 0 otherwise}

\item{seed}{optional, for reproducibility, only application when is.stoch = 1}
}
\value{
A data frame of model output
}
\description{
The main simulator which compiles and run an ODIN model in the background
}
\examples{
out <- meta_sim(N_pop = nrow(pop_zones),
                beta_i = 0.7,
                beta_v = 0.2,
                S0 = pop_zones$S0,
                I0 = pop_zones$I0,
                P0 = pop_zones$N,
                V0 = pop_zones$V0,
                m_weekday_day = m_weekday_day,
                m_weekend_day = m_weekend_day,
                m_weekday_night = m_weekday_night,
                m_weekend_night = m_weekend_night,
                delta_t = 0.5,
                tvac = vac_zones[, 1],
                vac_mat = as.matrix(vac_zones[, 2:7]),
                VtoS = 0.02,
                EtoIpresymp = 0.33,
                etopa = 0.5,
                pretoIsymp = 0.5,
                IasymptoR = 0.2,
                IsymptoRH = 0.5,
                istohr = 0.5,
                HtoRD = 0.33,
                htor = 0.5,
                RtoS = 0.02,
                vac_eff = 0.5,
                nsteps = 100,
                is.stoch = 0,
                seed = NULL)
}
