% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaODIN.R
\name{meta_sim}
\alias{meta_sim}
\title{Metapopulation respiratory virus model Simulator}
\usage{
meta_sim(
  N_pop,
  ts,
  tv,
  S0,
  I0,
  P0,
  V0,
  R0,
  m_weekday_day,
  m_weekday_night,
  m_weekend_day,
  m_weekend_night,
  delta_t,
  tvac,
  vac_mat,
  dv,
  de,
  pea,
  dp,
  da,
  ds,
  psr,
  dh,
  phr,
  dr,
  ve,
  nsteps,
  is.stoch = FALSE,
  seed = NULL
)
}
\arguments{
\item{N_pop}{Number of sub-populations}

\item{S0}{A vector of initial number of susceptible people in N_pop subpopulations}

\item{I0}{A vector of initial number of infected people in N_pop subpopulations}

\item{P0}{A vector of population sizes for N_pop subpopulations}

\item{V0}{A vector of initial number of vaccinated people in N_pop subpopulations}

\item{R0}{A vector of initial number of recovered people in N_pop subpopulations}

\item{m_weekday_day}{mixing matrix for weekday 6 am - 6 pm}

\item{m_weekday_night}{mixing matrix for weekday 6 pm - 6 am}

\item{m_weekend_day}{mixing matrix for weekend 6 am - 6 pm}

\item{m_weekend_night}{mixing matrix for weekend 6 pm - 6 am}

\item{delta_t}{A positive real number indicating the discrete time increment}

\item{tvac}{A vector of time indices for vaccination}

\item{vac_mat}{A matrix of order (length of tvac) x N_pop}

\item{dv}{mean number of days in Vaccinated state before waning immunity completely}

\item{de}{mean number of days in Exposed state}

\item{pea}{Proportion of people becoming infectious presymptomatic from exposed}

\item{dp}{mean number of days in Infectious presymptomatic state}

\item{da}{mean number of days in Infectious presymptomatic state}

\item{ds}{mean number of days in Infectious symptomatic state}

\item{psr}{Proportion of people recovered from infectious symptomatic}

\item{dh}{mean number of days in Hospitalized state}

\item{phr}{Proportion of people Recovered from Hospitalized}

\item{dr}{mean number of days in Recovered state}

\item{ve}{Vaccination efficacy, a value between 0 and 1}

\item{nsteps}{Number of discrete evolution in the simulation}

\item{is.stoch}{1 if the simulation is stochastic, 0 otherwise}

\item{seed}{optional, for reproducibility, only application when is.stoch = 1}

\item{beta_i}{Rate of transmission for susceptible, value between 0 and 1}

\item{beta_v}{Rate of transmission for vaccinated, value between 0 and 1}
}
\value{
A matrix of model output
}
\description{
The main simulator which compiles and run an ODIN model in the background
}
