% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaRVM-package.R
\docType{package}
\name{MetaRVM-package}
\alias{MetaRVM}
\alias{MetaRVM-package}
\title{MetaRVM: Meta-population Compartmental Model for Respiratory Virus Diseases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive R package for simulating respiratory virus epidemics using meta-population compartmental models. MetaRVM implements a stochastic SEIRD (Susceptible-Exposed-Infected-Recovered-Dead) framework with demographic stratification by age, race, and geographic zones. The package supports complex epidemiological scenarios including asymptomatic and presymptomatic transmission, hospitalization dynamics, vaccination campaigns, and time-varying contact patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://RESUME-Epi.github.io/MetaRVM/}
  \item \url{https://github.com/RESUME-Epi/MetaRVM}
  \item Report bugs at \url{https://github.com/RESUME-Epi/MetaRVM/issues}
}

}
\author{
\strong{Maintainer}: Arindam Fadikar \email{afadikar@anl.gov} (\href{https://orcid.org/0000-0001-7396-0350}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Charles Macal [contributor]
  \item Martinez Moyano Ignacio Javier [contributor]
  \item Ozik Jonathan [contributor]
}

}
\keyword{internal}
