% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_compartment}
\alias{plot_compartment}
\title{Title}
\usage{
plot_compartment(
  long_out_daily,
  conf_level,
  value_column,
  plot_type,
  color,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
